Move algorithm funcs to algorithm subpackage.
